+++ engine/runtime/src/main/scala/org/enso/compiler/core/IR.scala
@@ -299,8 +299,8 @@ object IR {
     * @param diagnostics compiler diagnostics for this node
     */
   @SerialVersionUID(
-    3692L
-  ) // prevents reading broken caches, see PR-3692 for details
+    3667L // removes Vector.Vector_Data constructor
+  )       // prevents reading broken caches, see PR-3692 for details
   sealed case class Module(
     imports: List[Module.Scope.Import],
     exports: List[Module.Scope.Export],